+++ lib/install_soa_monitoring.sh
@@ -26,8 +26,22 @@ git)
     test -d oci-tools || git clone https://github.com/rstyczynski/oci-tools.git
     ;;
 *)
+    if [ ! -f $tools_src/umc ]; then
+        echo "Error. umc not available at shared location. Put it there before proceeding"
+        exit 1
+    fi
     cp -rf $tools_src/umc ~/
+
+    if [ ! -f $tools_src/wls-tools ]; then
+        echo "Error. wls-tools not available at shared location. Put it there before proceeding"
+        exit 1
+    fi
     cp -rf $tools_src/wls-tools ~/
+    
+    if [ ! -f $tools_src/oci-tools ]; then
+        echo "Error. oci-tools not available at shared location. Put it there before proceeding"
+        exit 1
+    fi
     cp -rf $tools_src/oci-tools ~/
     ;;
 esac
@@ -44,44 +58,50 @@ export LC_ALL=en_US.UTF-8
 
 discoverWLS
 
-# Build SOA resource definition
-
-admin_host_protocol=$(getWLSjvmAttr ${wls_managed[0]} admin_host_protocol)
-admin_host_name=$(getWLSjvmAttr ${wls_managed[0]} admin_host_name)
-admin_host_port=$(getWLSjvmAttr ${wls_managed[0]} admin_host_port)
-
-adminURL_suffix=$admin_host_name:$admin_host_port
-admin_Server=${wls_admin[0]}
 
 if [ ! -z "$admin_Server" ]; then
-cat ~/umc/lib/soadms-probe.yaml | 
-sed "s/url: http:\/\/localhost:7001/url: $admin_host_protocol:\/\/$adminURL_suffix/" > ~/.umc/soadms-probe.yaml
+
+    # Build SOA resource definition
+
+    admin_host_protocol=$(getWLSjvmAttr ${wls_managed[0]} admin_host_protocol)
+    admin_host_name=$(getWLSjvmAttr ${wls_managed[0]} admin_host_name)
+    admin_host_port=$(getWLSjvmAttr ${wls_managed[0]} admin_host_port)
+
+    adminURL_suffix=$admin_host_name:$admin_host_port
+    admin_Server=${wls_admin[0]}
+
+
+    cat ~/umc/lib/soadms-probe.yaml | 
+    sed "s/url: http:\/\/localhost:7001/url: $admin_host_protocol:\/\/$adminURL_suffix/" > ~/.umc/soadms-probe.yaml
+
+    # set admin url
+    url=$admin_host_protocol://$adminURL_suffix
+
+    # Set WLS password
+    source ~/umc/bin/umc.h
+
+    pnp_vault save user$url $wls_user
+    pnp_vault save pass$url $wls_pass
+
+    # Test WLS dms/Spy connectivity
+
+    wls_user=$(pnp_vault read user$url)
+    wls_pass=$(pnp_vault read pass$url)
+
+    dms-collector --url $url --connect "$wls_user/$wls_pass" --count 1 --delay 1 --loginform --table soainfra_status
+
 else
-  rm -rf ~/.umc/soadms-probe.yaml
+    echo "Admin server not found. Test skipped."
 fi
 
-# Set WLS password
-
-source ~/umc/bin/umc.h
-
-url=$admin_host_protocol://$adminURL_suffix
-
-pnp_vault save user$url $wls_user
-pnp_vault save pass$url $wls_pass
-
-# Test WLS dms/Spy connectivity
-
-wls_user=$(pnp_vault read user$url)
-wls_pass=$(pnp_vault read pass$url)
-
-dms-collector --url $url --connect "$wls_user/$wls_pass" --count 1 --delay 1 --loginform --table soainfra_status
 
 # Start SOA collector
 
 if [ ! -z "$admin_Server" ]; then
     $HOME/umc/lib/soadms-service.sh soadms-probe.yaml restart
 else
-    echo "Admin server not found."
+    echo "Admin server not found. Stoping service as sanity step."
+    $HOME/umc/lib/soadms-service.sh soadms-probe.yaml  stop
 fi
 
 # init cron
@@ -90,19 +110,20 @@ cron_section_start="# START umc - $domain_name SOA DMS"
 cron_section_stop="# STOP umc - $domain_name SOA DMS"
 
 if [ ! -z "$admin_Server" ]; then
-cat >umc_dms.cron <<EOF
+    cat >umc_dms.cron <<EOF
 $cron_section_start
 1 0 * * * $HOME/umc/lib/soadms-service.sh soadms-probe.yaml restart
 $cron_section_stop
 EOF
 
-(crontab -l 2>/dev/null | 
-sed "/$cron_section_start/,/$cron_section_stop/d"
-cat umc_dms.cron) | crontab -
-rm umc_dms.cron
+    (crontab -l 2>/dev/null | 
+    sed "/$cron_section_start/,/$cron_section_stop/d"
+    cat umc_dms.cron) | crontab -
+    rm umc_dms.cron
 else
-(crontab -l 2>/dev/null | 
-sed "/$cron_section_start/,/$cron_section_stop/d") | crontab -
+    echo "Admin server not found. Removing entries from cron."
+    (crontab -l 2>/dev/null | 
+    sed "/$cron_section_start/,/$cron_section_stop/d") | crontab -
 fi
 
 crontab -l
