+++ lib/install_wls_monitoring.sh
@@ -18,8 +18,22 @@ git)
     test -d oci-tools || git clone https://github.com/rstyczynski/oci-tools.git
     ;;
 *)
+    if [ ! -f $tools_src/umc ]; then
+        echo "Error. umc not available at shared location. Put it there before proceeding"
+        exit 1
+    fi
     cp -rf $tools_src/umc ~/
+
+    if [ ! -f $tools_src/wls-tools ]; then
+        echo "Error. wls-tools not available at shared location. Put it there before proceeding"
+        exit 1
+    fi
     cp -rf $tools_src/wls-tools ~/
+    
+    if [ ! -f $tools_src/oci-tools ]; then
+        echo "Error. oci-tools not available at shared location. Put it there before proceeding"
+        exit 1
+    fi
     cp -rf $tools_src/oci-tools ~/
     ;;
 esac
@@ -86,7 +100,7 @@ fi
 if [ ! -z "$admin_Server" ]; then
     $HOME/umc/lib/wls-service.sh wls-probe.yaml restart
 else
-    echo "Admin server not found. Service start skipped. Stoping as sanity step."
+    echo "Admin server not found. Service start skipped. Stoping service as sanity step."
     $HOME/umc/lib/wls-service.sh wls-probe.yaml stop
 fi
